+++ src/plonk.rs
@@ -16,6 +16,7 @@ use crate::poly::{
 use crate::transcript::{ChallengeScalar, EncodedChallenge, Transcript};
 
 mod circuit;
+mod error;
 mod keygen;
 mod lookup;
 pub(crate) mod permutation;
@@ -25,6 +26,7 @@ mod prover;
 mod verifier;
 
 pub use circuit::*;
+pub use error::*;
 pub use keygen::*;
 pub use prover::*;
 pub use verifier::*;
@@ -131,33 +133,6 @@ pub struct ProvingKey<C: CurveAffine> {
     fixed_polys: Vec<Polynomial<C::Scalar, Coeff>>,
     fixed_cosets: Vec<Polynomial<C::Scalar, ExtendedLagrangeCoeff>>,
     permutation: permutation::ProvingKey<C>,
-}
-
-/// This is an error that could occur during proving or circuit synthesis.
-// TODO: these errors need to be cleaned up
-#[derive(Debug, PartialEq)]
-pub enum Error {
-    /// This is an error that can occur during synthesis of the circuit, for
-    /// example, when the witness is not present.
-    SynthesisError,
-    /// The structured reference string or the parameters are not compatible
-    /// with the circuit being synthesized.
-    IncompatibleParams,
-    /// The constraint system is not satisfied.
-    ConstraintSystemFailure,
-    /// Out of bounds index passed to a backend
-    BoundsFailure,
-    /// Opening error
-    OpeningError,
-    /// Transcript error
-    TranscriptError,
-    /// Instance provided has more rows than supported by circuit
-    NotEnoughRowsAvailable,
-    /// Instance provided exceeds number of available rows
-    InstanceTooLarge,
-    /// Circuit synthesis requires global constants, but circuit configuration did not
-    /// call [`ConstraintSystem::enable_constant`] on fixed columns with sufficient space.
-    NotEnoughColumnsForConstants,
 }
 
 impl<C: CurveAffine> ProvingKey<C> {